Rename symbols.
Implement a simple renaming header so that builds using our
local harfbuzz don't clash with other instances in the system.

This showed up with Qt (both 5 and 6) both ending up calling
our Harfbuzz library. This ends in tears due to our harfbuzz
expecting to call into an allocator that has a fz_context
available.

Thanks to https://github.com/DavidLazarescu for helping to track
this down.